+++ src/styles/search-agreement.scss
@@ -2,11 +2,6 @@
 
 @import '@navikt/ds-tokens/dist/tokens';
 
-.agreement-search-results {
-  border-top: 2px solid $a-border-divider;
-  padding-top: $a-spacing-2;
-}
-
 .agreement-search-results,
 .agreement-search-results ol,
 .search-results {
